Change Sample Size variable type
In ffmpeg Sample Size is 32-bit signed integer.

+++ mp4.cpp
@@ -430,7 +430,7 @@ void Mp4::repair(string filename) {
     size_t readFileResult; //resut of the file read operation
     uint32_t j = 0; //interleaving mask iterator (cyclic) or chunks counter (cyclic)
     uint32_t trackNumber = 0; //get it from interleaving mask
-    uint32_t minSmplLength; //get it from optMinSampleSize()
+    int32_t minSmplLength; //get it from optMinSampleSize()
     uint32_t samplesCount; //number of samples in current AU
     recoveredSample lengthResultsTmp; //temporary samples results per single chunk
     recoveredSample recoveredSTmp; //temporary storage of the recovered samples parameters
@@ -483,7 +483,7 @@ void Mp4::repair(string filename) {
             //split chunks into samples
             for (uint32_t i = 0; i < samplesCount; i++) {
 
-                minSmplLength = track.optMinSampleSize(track, brokenFile, offset, track.codec.codecParamByFFmpeg);
+                minSmplLength = (int32_t) track.optMinSampleSize(track, brokenFile, offset, track.codec.codecParamByFFmpeg);
                 //AAC minSampleLength = 6;
 
                 logMe(LOG_DBG, "");
